Move goals tests to overview folder
