+++ integration_tests/e2e/overview/goal/archiveGoal.cy.ts
@@ -1,12 +1,12 @@
-import Page from '../../pages/page'
-import ArchiveGoalPage from '../../pages/goal/ArchiveGoalPage'
-import ReasonToArchiveGoalValue from '../../../server/enums/ReasonToArchiveGoalValue'
-import ReviewArchiveGoalPage from '../../pages/goal/ReviewArchiveGoalPage'
-import OverviewPage from '../../pages/overview/OverviewPage'
-import { putRequestedFor } from '../../mockApis/wiremock/requestPatternBuilder'
-import { urlEqualTo } from '../../mockApis/wiremock/matchers/url'
-import { matchingJsonPath } from '../../mockApis/wiremock/matchers/content'
-import GoalsPage from '../../pages/goal/GoalsPage'
+import Page from '../../../pages/page'
+import ArchiveGoalPage from '../../../pages/goal/ArchiveGoalPage'
+import ReasonToArchiveGoalValue from '../../../../server/enums/ReasonToArchiveGoalValue'
+import ReviewArchiveGoalPage from '../../../pages/goal/ReviewArchiveGoalPage'
+import OverviewPage from '../../../pages/overview/OverviewPage'
+import { putRequestedFor } from '../../../mockApis/wiremock/requestPatternBuilder'
+import { urlEqualTo } from '../../../mockApis/wiremock/matchers/url'
+import { matchingJsonPath } from '../../../mockApis/wiremock/matchers/content'
+import GoalsPage from '../../../pages/goal/GoalsPage'
 
 context('Archive a goal', () => {
   const prisonNumber = 'G6115VJ'
