+++ integration_tests/e2e/overview/goal/unarchiveGoal.cy.ts
@@ -1,12 +1,12 @@
-import Page from '../../pages/page'
-import UnarchiveGoalPage from '../../pages/goal/UnarchiveGoalPage'
-import OverviewPage from '../../pages/overview/OverviewPage'
-import { urlEqualTo } from '../../mockApis/wiremock/matchers/url'
-import { putRequestedFor } from '../../mockApis/wiremock/requestPatternBuilder'
-import { matchingJsonPath } from '../../mockApis/wiremock/matchers/content'
-import AuthorisationErrorPage from '../../pages/authorisationError'
-import ViewArchivedGoalsPage from '../../pages/goal/ViewArchivedGoalsPage'
-import GoalsPage from '../../pages/goal/GoalsPage'
+import Page from '../../../pages/page'
+import UnarchiveGoalPage from '../../../pages/goal/UnarchiveGoalPage'
+import OverviewPage from '../../../pages/overview/OverviewPage'
+import { urlEqualTo } from '../../../mockApis/wiremock/matchers/url'
+import { putRequestedFor } from '../../../mockApis/wiremock/requestPatternBuilder'
+import { matchingJsonPath } from '../../../mockApis/wiremock/matchers/content'
+import AuthorisationErrorPage from '../../../pages/authorisationError'
+import ViewArchivedGoalsPage from '../../../pages/goal/ViewArchivedGoalsPage'
+import GoalsPage from '../../../pages/goal/GoalsPage'
 
 context('Unarchive a goal', () => {
   const prisonNumber = 'G6115VJ'
