+++ integration_tests/e2e/overview/goal/updateGoal.cy.ts
@@ -1,14 +1,14 @@
-import Page from '../../pages/page'
-import UpdateGoalPage from '../../pages/goal/UpdateGoalPage'
-import OverviewPage from '../../pages/overview/OverviewPage'
-import ReviewUpdateGoalPage from '../../pages/goal/ReviewUpdateGoalPage'
-import AuthorisationErrorPage from '../../pages/authorisationError'
-import Error404Page from '../../pages/error404'
-import Error500Page from '../../pages/error500'
-import { putRequestedFor } from '../../mockApis/wiremock/requestPatternBuilder'
-import { urlEqualTo } from '../../mockApis/wiremock/matchers/url'
-import { matchingJsonPath } from '../../mockApis/wiremock/matchers/content'
-import GoalsPage from '../../pages/goal/GoalsPage'
+import Page from '../../../pages/page'
+import UpdateGoalPage from '../../../pages/goal/UpdateGoalPage'
+import OverviewPage from '../../../pages/overview/OverviewPage'
+import ReviewUpdateGoalPage from '../../../pages/goal/ReviewUpdateGoalPage'
+import AuthorisationErrorPage from '../../../pages/authorisationError'
+import Error404Page from '../../../pages/error404'
+import Error500Page from '../../../pages/error500'
+import { putRequestedFor } from '../../../mockApis/wiremock/requestPatternBuilder'
+import { urlEqualTo } from '../../../mockApis/wiremock/matchers/url'
+import { matchingJsonPath } from '../../../mockApis/wiremock/matchers/content'
+import GoalsPage from '../../../pages/goal/GoalsPage'
 
 context('Update a goal', () => {
   beforeEach(() => {
